Fix problem where the wordpress sync plugin was overwriting woocommerce orders
We were setting the current user for permissions purposes when exporting lesson content managed by other teachers. However, we set the current user before we checked to see if the post that triggered the action was even a valid one (hint: orders are not). We also changed the user but did not change back to the original user before returning. This culminated in all orders on stripe showing up under the default user.